* 'master' of https://github.com/craigrmccown/cs-3210:
  replication functional when node is added
  fixed bugs
  node topology consistent across nodes
  bugfixes
  topology additions working

+++ project3/slave_node/register.py
@@ -9,18 +9,19 @@ mongo = MongoClient('127.0.0.1', 27017)
 
 
 def register():
-    registration = requests.post(''.join(['http://', master_ip_address, ':', master_port_number, '/register']))
+    registration = requests.post(''.join(['http://', master_ip_address, ':', str(master_port_number), '/register']))
     assert registration.status_code == 200
 
     response_body = registration.json()
     slave_node_id = response_body.get('node_id')
     topology = response_body.get('topology')
 
-    db = mongo['rpfs_slave_db_' + str(slave_node_id)]
-    db.topology.drop()
+    if len(topology) > 0:
+        db = mongo['rpfs_slave_db_' + str(slave_node_id)]
+        db.topology.drop()
 
-    for node in topology:
-        db.topology.insert(node)
+        for node in topology:
+            db.topology.insert(node)
 
     return slave_node_id
